fix(enchanter): migrate 26.1.2 enchantment APIs

+++ Fabric/src/main/java/vazkii/botania/fabric/mixin/EnchantmentMenuFabricMixin.java
@@ -34,13 +34,13 @@ public class EnchantmentMenuFabricMixin {
 		for (int x = -1; x <= 1; ++x) {
 			for (int z = -1; z <= 1; ++z) {
 				if ((x != 0 || z != 0) && level.isEmptyBlock(pos.offset(x, 0, z)) && level.isEmptyBlock(pos.offset(x, 1, z))) {
-					obj += getPylonValue(level.getBlockState(pos.offset(x * 2, 0, z * 2)), level, pos);
-					obj += getPylonValue(level.getBlockState(pos.offset(x * 2, 1, z * 2)), level, pos);
+					obj += (int) getPylonValue(level.getBlockState(pos.offset(x * 2, 0, z * 2)), level, pos);
+					obj += (int) getPylonValue(level.getBlockState(pos.offset(x * 2, 1, z * 2)), level, pos);
 					if (x != 0 && z != 0) {
-						obj += getPylonValue(level.getBlockState(pos.offset(x * 2, 0, z)), level, pos);
-						obj += getPylonValue(level.getBlockState(pos.offset(x * 2, 1, z)), level, pos);
-						obj += getPylonValue(level.getBlockState(pos.offset(x, 0, z * 2)), level, pos);
-						obj += getPylonValue(level.getBlockState(pos.offset(x, 1, z * 2)), level, pos);
+						obj += (int) getPylonValue(level.getBlockState(pos.offset(x * 2, 0, z)), level, pos);
+						obj += (int) getPylonValue(level.getBlockState(pos.offset(x * 2, 1, z)), level, pos);
+						obj += (int) getPylonValue(level.getBlockState(pos.offset(x, 0, z * 2)), level, pos);
+						obj += (int) getPylonValue(level.getBlockState(pos.offset(x, 1, z * 2)), level, pos);
 					}
 				}
 			}
